Performance documento fiscal e ajustes front end

+++ src/main/java/br/com/pi/projeto_RD/model/dto/ItensDfDTO.java
@@ -3,6 +3,7 @@ package br.com.pi.projeto_RD.model.dto;
 import br.com.pi.projeto_RD.model.entity.ProdutoFilialEstoqueEntity;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -13,8 +14,8 @@ public class ItensDfDTO {
     private BigInteger cdProduto;
     private String nmProduto;
     private Integer qtItem;
-    private Double vlItem;
-    private Double pcIcms;
-    private Double vlIcms;
+    private BigDecimal vlItem;
+    private BigDecimal pcIcms;
+    private BigDecimal vlIcms;
 
 }
